Rename MockData provider
