+++ src/app/app.module.ts
@@ -12,7 +12,7 @@ import { ProgressPage } from '../pages/progress/progress';
 import { FriendsPage } from '../pages/friends/friends';
 import { ProfilePage } from '../pages/profile/profile';
 import { ExperiencesPage } from '../pages/experiences/experiences';
-import { DataProvider } from '../providers/mock-data/mock-data';
+import { DataProvider } from '../providers/data/data';
 import { QuestionPage } from "../pages/question/question";
 import { ActivityPage } from '../pages/activity/activity';
